requirements.txt
dill ~= 0.3
numpy ~= 1.26
pillow ~= 10.2